+++ src/module/chatMesssageHandling.ts
@@ -371,11 +371,23 @@ export let hideStuffHandler = (message, html, data) => {
       html.find(".midi-qol-other-roll").find(".dice-formula").remove()
       html.find(".midi-qol-bonus-roll").find(".dice-tooltip").remove()
       html.find(".midi-qol-bonus-roll").find(".dice-formula").remove()
-/*
+/* TODO remove this pending feedback
       html.find(".midi-qol-damge-roll").find(".dice-roll").replaceWith(`<span>${i18n("midi-qol.DiceRolled")}</span>`);
       html.find(".midi-qol-other-roll").find(".dice-roll").replaceWith(`<span>${i18n("midi-qol.DiceRolled")}</span>`);
       html.find(".midi-qol-bonus-roll").find(".dice-roll").replaceWith(`<span>${i18n("midi-qol.DiceRolled")}</span>`);
 */
+    } else if (d20AttackRoll && configSettings.hideRollDetails === "hitDamage") {
+      const hitFlag = getProperty(message.data.flags, "midi-qol.isHit");
+        const hitString = hitFlag === undefined ? "" : hitFlag ? i18n("midi-qol.hits") : i18n("midi-qol.misses");
+        html.find(".midi-qol-attack-roll .dice-total").text(`${hitString}`);
+        html.find(".dice-tooltip").remove();
+        html.find(".dice-formula").remove();
+        html.find(".midi-qol-damage-roll").find(".dice-tooltip").remove()
+        html.find(".midi-qol-damage-roll").find(".dice-formula").remove()
+        html.find(".midi-qol-other-roll").find(".dice-tooltip").remove()
+        html.find(".midi-qol-other-roll").find(".dice-formula").remove()
+        html.find(".midi-qol-bonus-roll").find(".dice-tooltip").remove()
+        html.find(".midi-qol-bonus-roll").find(".dice-formula").remove()
     } else if (configSettings.hideRollDetails === "all" || message.data.blind) {
       html.find(".dice-roll").replaceWith(`<span>${i18n("midi-qol.DiceRolled")}</span>`);
       //TODO this should probably just check formula